requirements.txt
PyYAML
MySQL-python
fake-factory
